feat(finetune.sh) - skip checkpoint download

+++ finetune.sh
@@ -5,7 +5,9 @@ export PYTHONPATH=$(pwd):$PYTHONPATH
 
 # Download checkpoint and inference
 CHECKPOINT_PATH="CRNN_note_F1=0.9677_pedal_F1=0.9186.pth"
-wget -O "$CHECKPOINT_PATH" "https://zenodo.org/record/4034264/files/CRNN_note_F1=0.9677_pedal_F1=0.9186.pth?download=1"
+if [ ! -f "$CHECKPOINT_PATH" ]; then
+    wget -O "$CHECKPOINT_PATH" "https://zenodo.org/record/4034264/files/CRNN_note_F1=0.9677_pedal_F1=0.9186.pth?download=1"
+fi
 
 # Note and pedal checkpoints
 NOTE_CHECKPOINT="onset_offset_frame_velocity_pretrained.pth"
